+++ source/js/inverter/inverter.js
@@ -134,9 +134,12 @@ define(
   }
 
   function reverseElems($elems, doReverse) {
-    $elems.each(function() {
-      var src = this.getAttribute('src'),
-        reverseElem = (src && SHOULD_USE_PROXY) ? reverseElemProxy : reverseElemCss;
+    $elems.each(function () {
+      var
+        src = this.getAttribute('src'),
+        // The image proxy can't handle svg images
+        isSVG = imgClassifier.isSVGSource(src),
+        reverseElem = (src && !isSVG && SHOULD_USE_PROXY) ? reverseElemProxy : reverseElemCss;
       reverseElem($(this), doReverse, src);
     });
   }